+++ _docs/pages/docs.tsx
@@ -20,8 +20,12 @@ const docs: Doc[] = [
     path: "expect",
   },
   {
-    name: "fake",
-    path: "fake",
+    name: "dummy",
+    path: "dummy",
+  },
+  {
+    name: "mock",
+    path: "mock",
   },
 ];