+++ matcher/to_have_been_called_times.ts
@@ -1,9 +1,12 @@
 // Copyright 2021-Present the Unitest authors. All rights reserved. MIT license.
 // This module is browser compatible.
-import type { Mock } from "../mock/types.ts";
+import type { MockObject } from "../mock/mock.ts";
 import type { MatchResult } from "./types.ts";
 
-function toHaveBeenCalledTimes({ mock }: Mock, expected: number): MatchResult {
+function toHaveBeenCalledTimes(
+  { mock }: MockObject,
+  expected: number,
+): MatchResult {
   return {
     pass: mock.calls.length === expected,
     expected,