+++ matcher/to_have_last_returned_with.ts
@@ -3,9 +3,12 @@
 import { equal } from "../helper/equal.ts";
 import type { MatchResult } from "./types.ts";
 import { takeLast } from "./utils.ts";
-import type { Mock, MockResult } from "../mock/types.ts";
+import type { MockObject, MockResult } from "../mock/mock.ts";
 
-function predict(mockResults: MockResult[], expected: unknown): boolean {
+function predict(
+  mockResults: readonly MockResult[],
+  expected: unknown,
+): boolean {
   const { type, value } = takeLast(1, mockResults)[0] ?? {};
 
   return type === "return" && equal(value, expected);
@@ -13,7 +16,7 @@ function predict(mockResults: MockResult[], expected: unknown): boolean {
 
 /** Use to test the specific value that a mock function last returned. */
 function toHaveLastReturnedWith(
-  { mock }: Mock,
+  { mock }: MockObject,
   expected: unknown,
 ): MatchResult {
   return {