+++ matcher/to_have_returned_with_test.ts
@@ -1,7 +1,7 @@
 // Copyright 2021-Present the Unitest authors. All rights reserved. MIT license.
 import { predict, toHaveReturnedWith } from "./to_have_returned_with.ts";
 import { assertEquals, assertFail, assertSuccess } from "../dev_deps.ts";
-import type { MockResult } from "../mock/types.ts";
+import type { MockResult } from "../mock/mock.ts";
 import { fn } from "../mock/fn.ts";
 
 Deno.test({