+++ mock/mod.ts
@@ -1,4 +1,5 @@
 // Copyright 2021-Present the Unitest authors. All rights reserved. MIT license.
 // This module is browser compatible.
-export { fn } from "./fn.ts";
-export * from "./types.ts";
+export * from "./fn.ts";
+export * from "./mock.ts";
+export * from "./spy_on.ts";